\name{print.gevreg}
\alias{print.gevreg}
\title{Print output of a gevreg object.}
\usage{
\method{print}{gevreg}(x, digits = max(3, getOption("digits") - 3), ...)
}
\arguments{
\item{x}{gevreg object.}

\item{digits}{Number of digits.}
\item{...}{Additional parameters.}
}
\description{
Print output of a gevreg object.
}
